pyproject.toml
[project]
name = "cityevents"
version = "0.1.0"
description = ""
authors = [
    {name = "etadevosyan"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aiogram",
    "fastapi",
    "pydantic>=2.10,<3.0",
    "sqlalchemy==2.0.35",
    "uvicorn==0.31.0",
    "jinja2==3.1.4",
    "pydantic-settings==2.5.2",
    "alembic==1.13.3",
    "aiosqlite==0.20.0",
    "greenlet",
    "pydantic-ai==1.0.1",
    "pyyaml>=6.0.0,<7.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

